+++ Events/JumpedEvent.cs
@@ -84,26 +84,34 @@ namespace EddiEvents
         [PublicAPI("The government of the system to which the commander has jumped")]
         public string government => (controllingfaction?.Government ?? Government.None).localizedName;
 
-        // Powerplay properties (only when pledged)
+        // Powerplay properties
 
-        [PublicAPI( "(Only when pledged) The localized powerplay power controlling the star system, if any. If the star system is `Contested`, this will be empty" )]
+        [PublicAPI( "The localized powerplay power controlling the star system, if any.  If the star system is `Unoccupied` or `Contested` then this will be empty." )]
         public string power => ( Power ?? Power.None ).localizedName;
 
-        [PublicAPI( "(Only when pledged) The localized names of powerplay powers contesting control of the star system, if any" )]
+        [PublicAPI( "The localized names of powerplay powers having star systems with acquisition radii which overlap the star system, less the controlling power, if any" )]
+        public List<string> nearbypowers => NearbyPowers?
+            .Select( p => p.localizedName )
+            .ToList();
+
+        [PublicAPI( "The localized names of powerplay powers contesting control of an uncontrolled star system, if any" )]
         public List<string> contestingpowers => ContestingPowers?
             .Select( p => p.localizedName )
             .ToList();
 
-        [PublicAPI( "(Only when pledged) The state of powerplay efforts within the star system" )]
+        [PublicAPI( "The localized state of powerplay efforts within the star system" )]
         public string powerstate => ( PowerState ?? PowerplayState.Unoccupied ).localizedName;
 
-        [PublicAPI( "(Only when pledged) The powerplay power controlling the star system, if any, as an object. If the star system is `Contested`, this will be empty" )]
+        [PublicAPI( "The powerplay power controlling the star system, if any, as an object." )]
         public Power Power { get; private set; }
 
-        [PublicAPI( "(Only when pledged) Powerplay powers contesting control of the star system, if any, as objects" )]
+        [PublicAPI( "Powerplay powers having star systems with acquisition radii which overlap the star system, less the controlling power, if any, as objects" )]
+        public List<Power> NearbyPowers { get; set; }
+
+        [PublicAPI( "Powerplay powers contesting control of an uncontrolled star system, if any, as objects" )]
         public List<Power> ContestingPowers { get; set; }
 
-        [PublicAPI( "(Only when pledged) The state of powerplay efforts within the star system, as an object" )]
+        [PublicAPI( "The state of powerplay efforts within the star system, as an object" )]
         public PowerplayState PowerState { get; private set; }
 
         // Thargoid War
@@ -142,12 +150,12 @@ namespace EddiEvents
             this.controllingfaction = controllingfaction;
             this.factions = factions;
             this.conflicts = conflicts;
-            this.Economy = (economy ?? Economy.None);
-            this.Economy2 = (economy2 ?? Economy.None);
-            this.securityLevel = (security ?? SecurityLevel.None);
+            this.Economy = ( economy ?? Economy.None );
+            this.Economy2 = ( economy2 ?? Economy.None );
+            this.securityLevel = ( security ?? SecurityLevel.None );
             this.population = population;
             this.Power = controllingPower;
-            this.ContestingPowers = powerplayPowers?
+            this.NearbyPowers = powerplayPowers?
                 .Where( p => p.edname != Power?.edname )
                 .ToList();
             this.PowerState = powerplayState;